Add script to update database to Unaligned Expansion
This script updates the database keeping the same ids as the previous
database. New cards were added which change the ids of previous cards.
Existent decks and favorite cards may not point to the correct ids when
the database is updated. This script remaps ids which were changed to be
the same as they were in the previous database.