Improved Dialog response when directory contains a lot of files
Now the listing of directories is done in another thread to not block the UI thread. This improves touch response when the directory has a lot of files. As is the case when selecting the folder which contains the images since they are a lot.

Also implemented onSaveInstanceState to save the current being selected folder. This way it isn't lost on screen rotation or when the user changes to another app.