Fix a problem with shared element transition
This layout uses imageviews in two rows. But they aren't always shown. When the last one isn't shown (it only appears if the card has 7 disciplines, which is very rare), the path of the transition animation is incorrectly computed and would result in a little bump when the transition finishes. By removing this last imageview from the layout, the transition is computed correctly as the layout stays with a stable size.