Improve card image loading
Cards with accented chars names are properly recognized and loaded.

Improved Utils.loadCardImage to not be responsible to derivate the cardfilename. Now it accepts a filename instead. Also removed the decision making of which placeholder card (gold back or green back) to be shown in case the card isn't found. This decision is made by the caller and passed to loadCardImage as a resource id.