Add a lot of support methods and improvements
Add method to load thumbnails.

Add method to get an activity from a view.

Add callback support which is called the the BitmapDrawable is loaded. This helps the label coloring support to card details based on colors of the image loaded.

Remove need to receive activity and resources object to load images.