Fix searchbar layout problem when using API 10
When running on api 10, the drawer appears over the edittext instead of keeping on its left as usual. After setting the layout param to use all the width of the parent (MATCH_PARENT), the layout was being drawn correctly.

When adding the customview to the toolbar, although it is specified match_parent as the width, it seems that this is changed to wrap content. This is confirmed by using the layout inspector where you can see the layoutparam of wrap_content.