Fix a bug where an imageview could have a reference to a recycled bitmap
If the cache was cleared (maybe because of a low memory situation) those imageviews will have references to BitmapDrawables which have references to recycled bitmaps. This happened when those BitmapDrawables were evicted from the cache.

In this case, clear the imageViews drawables (BitmapDrawable) references. This is needed because we are loading the bitmap drawables using asynctask in a backgrond thread. This means the UI thread may try to draw the imageViews (which have the recycled bitmap reference) before we can load the new BitmapDrawables crashing the application.