Filter only the list affected by the filtering option being selected by the user
Only the name filter will be applied to both lists. Other filters will be only applied to the specific list.

Performance is improved because the list which isn't affected by the filter isn't reloaded unnecessarily.

Also, the delay for processing filters is only applied to the card name filtering where we have to wait the user to stop typing before applying.

Fix #30