Add observable for searchTextHint and needRefreshFlag
Instead of indirectly inferring if data needs to be refreshed by subscribing to preference changes observables, now CardsViewModel exposes an observable which emits an item to indicate that a referesh is needed.

Add an observable to specify the text hint of the searchbar depending on the preference to include the card text or not.

Both additions remove logic from the view and put it into the modelview allowing the view to simply react accordingly.