Improve library card details screen
Library card details screen had a lot of cardviews. Refactored it to contain only two cardviews: one for the text and another for other information. This makes working with the screen easier by not having to use many cardviews and improves clarity of the screen by putting all the information in a more condensed form.