Improve fix for issue #28
After changing cards listing update to send animated modifications with dispatchUpdatesTo(Adapter) in https://github.com/franciscojunior/Vampidroid/commit/fe3021291244e2b35a9fa2eeab3e180c5d7265c8 when the cards images folder is changed, this isn't reflected in the listing.
This is caused by the fact that the listing isn't refreshed anymore. And so, the images aren't reloaded. Now, when the folder is changed, force a full refresh so the images are reloaded from the new folder.
In order to do that, CardsViewModel now makes available an Observable (getNeedRefreshCardImages()) which emits whenever this folder is changed so the application can react accordingly.

The previous getNeedRefreshFlag Observable which was also used to emit whenever both the folder was changed and the preference for search inside the text card was changed was replaced by the getNeedRefreshCardsListing which only emits when the preference is changed and the cards listing needs to be refreshed. Not related to the images folder.

Add tests for those Observables