Avoid unnecessary emission of the refreshcardimages observable
When subscribing to the refreshcardimages observable, it was emitting an item. But it was not needed because at the time it is subscribed, the activity is being created which will load all the cards anyway.
So, skip the first emission to avoid loading the cards twice when starting the application.