Add Architecture Components Viemodel support
Changed all view models to be Architecture Components Viewmodels.

With this change, we can better manage the state of the application by
using the Android support to keep view models in sync with the
activities lifecycle.

The comunication between activities and fragments is much better. For
example, the SettingsViewModel allowed the removal of a communication
interface between two fragments because both fragments could talk to the
same view model instance.

Fix #48