Use AsyncLayoutInflater in CardFilters widget
CardFilters widget uses a complex layout which was slowing down app
initialization, restore from background and screen rotation. In other
words, anything which caused its views to be inflated.

In order to mitigate that, now CardFilters widget inflates
asynchronously its layout so the UI thread is free.

This required a lot of changes regarding initialization of the widget as
well as saving and restoring its state. As the layout would be inflated
later, a lot of initialization code was migrated to the callback of the
layout inflation.

Thanks guys of Android team for providing such a nice utility class!

Fix #51, #52