Use recyclerview.extensions.ListAdapter base class
With support library 27.1.0, ListAdapter base class was made available.
This class makes it easy to create list diffs in a background thread and
dispatches those changes to the recyclerview creating the animations.

Now, instead of compute by hand the diffs with DiffUtil, use this base
class which does all the heavy lift.

Update support library to 27.1.0