Add support to load card images from remote server
Since the beginning, Vampidroid only used card images stored in a local folder. This patch adds the ability to load cards from a remote server. Just specify the server path of the cards in the settings and Vampidroid will load it on demand.

Thanks to Glide, this change was very easy to implement.

Fix #60